// src/app/domain/model/get-game-details200-response.ts
/**
 * Wordle API
 * API para jugar Wordle, un juego de adivinanzas de palabras de 5 letras.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GetGameDetails200ResponseGuessesInner } from './get-game-details200-response-guesses-inner';


export interface GetGameDetails200Response { 
    /**
     * ID único del juego
     */
    gameId?: string;
    /**
     * Estado actual del juego -> 0(En progreso), 1(Victoria), 2(Derrota)
     */
    status?: GetGameDetails200Response.StatusEnum;
    /**
     * Número de intentos restantes
     */
    attemptsLeft?: number;
    /**
     * Palabra que se debe adivinar
     */
    wordToGuess?: string;
    guesses?: Array<GetGameDetails200ResponseGuessesInner>;
}
export namespace GetGameDetails200Response {
    export type StatusEnum = '0' | '1' | '2';
    export const StatusEnum = {
        _0: '0' as StatusEnum,
        _1: '1' as StatusEnum,
        _2: '2' as StatusEnum
    };
}