// src/app/domain/model/make-guess200-response.ts
/**
 * Wordle API
 * API para jugar Wordle, un juego de adivinanzas de palabras de 5 letras.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MakeGuess200Response { 
    /**
     * Resultado del intento (por ejemplo, \"2 posición y letra correcta, 1 posición incorrecta y letra correcta, 0 letra no existe en la palabra\")
     */
    result?: string;
    /**
     * Número de intentos restantes
     */
    attemptsLeft?: number;
    /**
     * Indica si el juego ha sido ganado
     */
    isGameWon?: boolean;
}