// src/app/domain/model/start-game400-response.ts
/**
 * Wordle API
 * API para jugar Wordle, un juego de adivinanzas de palabras de 5 letras.
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface StartGame400Response { 
    /**
     * Descripción del error
     */
    error?: string;
}